requirements.txt
fastapi
uvicorn
gunicorn
fasttext-wheel
wget
numpy<2.0.0
scikit-learn